modificacoes nas etiquetas

+++ src/main/java/com/example/teoloblog/service/funcao/FuncaoServiceImpl.java
@@ -2,9 +2,11 @@ package com.example.teoloblog.service.funcao;
 
 import com.example.teoloblog.convert.funcao.FuncaoConvert;
 import com.example.teoloblog.domain.funcao.Funcao;
+import com.example.teoloblog.domain.usuario.Usuario;
 import com.example.teoloblog.dto.funcao.FuncaoDTO;
 import com.example.teoloblog.dto.funcao.FuncaoFormDTO;
 import com.example.teoloblog.repository.funcao.FuncaoRepository;
+import com.example.teoloblog.repository.usuario.UsuarioRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -17,6 +19,8 @@ import java.util.Optional;
 public class FuncaoServiceImpl implements FuncaoService {
 
     private final FuncaoRepository funcaoRepository;
+    private final UsuarioRepository usuarioRepository;
+
     @Override
     public List<FuncaoDTO> listarFuncoes() {
         List<Funcao> funcoes = funcaoRepository.findAll();
@@ -27,7 +31,6 @@ public class FuncaoServiceImpl implements FuncaoService {
     public FuncaoDTO buscarFuncaoPorCodigo(Integer codigo) throws Exception{
         Optional<Funcao> funcaoOpt = funcaoRepository.findById(codigo);
         if(!funcaoOpt.isPresent()){
-            throw new Exception("Não existe essa função");
         }
         return FuncaoConvert.funcaoDomainToDTO(funcaoOpt.get());
     }
@@ -54,10 +57,17 @@ public class FuncaoServiceImpl implements FuncaoService {
 
      @Override
     public void deletaFuncao(Integer codigo) throws Exception {
+
         Optional<Funcao> funcaoOpt = funcaoRepository.findById(codigo);
         if(!funcaoOpt.isPresent()){
-            throw new Exception("Função não encotrada");
+            throw new Exception("Função não encontrada");
         }
+
+        Optional<Usuario> usuarioOpt = usuarioRepository.findByFuncao(funcaoOpt.get());
+        if(usuarioOpt.isPresent()){
+            throw new Exception("Essa função está sendo utilizada em um usuário");
+        }
+
         funcaoRepository.delete(funcaoOpt.get());
     }
 }
